requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
requests==2.31.0
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.0.0
numpy==1.24.3
